+++ sources/Struct.public_methods_definitions.c
@@ -1,3 +1,4 @@
+#include <stddef.h>           /* size_t ptrdiff_t */
 #include "oyranos_cmm.h" /* oyObjectInfoStatic_s */
 
 const char * (*oyStruct_GetTextFromModule_p) (
@@ -196,7 +197,7 @@ int          oyStruct_CheckType      ( oyStruct_s        * obj,
       } else
         return 2;
 
-      if((uint32_t)inheritance < (uint32_t)oyOBJECT_MAX)
+      if((ptrdiff_t)inheritance < (ptrdiff_t)oyOBJECT_MAX)
       {
         WARNc1_S( "non plausible inheritance pointer: %s", 
                   oyStruct_GetInfo(obj,0) );
